Update families.csv - Wordspace
Hey Marc, the page stopped refreshing as I endeted data after an  hour or so of steady entry so I thought might be a sign. But also this non refreshing meant I could just keep entering values without scrolling ( or without much scrolling after a while ) which was WAY more efficient. I wonder if there is a way to hold off on refreshing the page. I did hit on the idea of limiting the search by a bigram aa oo etc because that also meant less scrolling. But this was before the page stopped refreshing with entry of a value.